requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
httpx[http2]==0.27.0
sse-starlette==2.1.3
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
pytest==8.2.1
pytest-asyncio==0.23.6
ruff==0.4.5
prometheus_client==0.19.0
pylint==4.0.4
